Enable multiple SSL connections by using AXTLS fd
axTLS lets us pass in a "file descriptor" which is used only by our own
code and could be anything that fits into an int.  We were passing in 0
all the time which means axtls would get confused if there were
multiple simultaneous connections.

Instead of 0, pass in (this), a pointer to our object, where we now have
a normal class variable io_ctx, in place of the single static s_io_ctx.
By doing so we guarantee multiple connections get unique multiple FDs,
and remove the need for any static tracking of the io_ctx array/map.