Add ESP8266WebServerSecure, a simple HTTPS server
Add support for simple, easy-to-use WebServer using the SecureServer.
For the end user, operates the same as the existing class with the
addition of a key/certificate setting stage (and using port 443, of
course).  Sample code included.

Due to some ugly C++ object slicing issues, ended up duplicating
a few functions that used the assignment operator to new local
WiFiClients.  Only the client communications bits needed this treatment
and the actual unique secure server class is quite small and simple.

Also found and fixed a memory leak in WiFiClientSecure in this
use case:  Unreference (and potentially delete) and clear the
WiFiClientSecure's SSLContext on a ::stop().  The destructor for
WiFiClientSecure sometimes isn't called when the WiFiClientSecure is
cast to a WiFiClient (object slicing).  This causes SSLContext* _ssl
not to be freed, rapidly filling up memory and dying after 4-5
connections

To avoid this, do the dereference and clear in the ::stop() which
is safe since no more operations on the WiFiClientSecure can occur
after the connection is ::stopped().